feat(publish-release): add output to job
This enables users of the workflow to chain other jobs conditionally.

The specific use case for this is in our Zapier integration. We use
publish-release workflow to manage a GitHub release and changelog and if the job
succeeds and has published a package, we want to trigger a Zapier deployment.